broke up json/binary exception handling
